+++ Sources/EvolutionMetadataExtraction/Utilities/Networking.swift
@@ -88,7 +88,9 @@ struct GitHubPullFileItem: Codable {
     }
 
     func proposalSpec(project: Project, sortIndex: Int) ->  ProposalSpec {
-        ProposalSpec(project: project, url: URL(string: raw_url)!, sha: sha, sortIndex: sortIndex)
+        // The url values contain percent-encoded slashes (%2F) in directory paths
+        // This prevented proper extraction of the proposal number from the proposal filename
+        return ProposalSpec(project: project, url: URL(string: raw_url.removingPercentEncoding!)!, sha: sha, sortIndex: sortIndex)
     }
 
 }